+++ src/main/java/online/hatsunemiku/tachideskvaadinui/services/TrackingCommunicationService.java
@@ -34,7 +34,7 @@ public class TrackingCommunicationService {
    * @param onlyWhenBigger a flag indicating whether to update the progress only when the chapter is
    *     bigger than the current progress
    */
-  public void setChapterProgress(int mangaId, int chapter, boolean onlyWhenBigger) {
+  public void setChapterProgress(int mangaId, float chapter, boolean onlyWhenBigger) {
     var tracker = dataService.getTracker(mangaId);
 
     if (onlyWhenBigger) {
